requirements.txt
asgiref==3.5.0
astroid==2.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-dataclasses==1.1.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
flake8-commas==2.1.0
flake8-polyfill==1.0.2
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
numpy==1.22.3
parameterized==0.8.1
pep8-naming==0.12.1
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.13.5
pytz==2022.1
pyzmq==24.0.1
requests==2.28.1
scipy==1.8.0
sqlparse==0.4.2
StochPy==2.4
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
